GUI file architecture restruct
