+++ frontend/intruder.py
@@ -1,0 +1,7 @@
+import customtkinter as ctk
+
+
+class GUIIntruder(ctk.CTkFrame):
+    def __init__(self, master, root):
+        super().__init__(master)
+        self.configure(fg_color="transparent")
